pyproject.toml
[project]
name = "culinary-canvas"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.5.3",
    "daphne>=4.2.1",
    "django>=5.2.6",
    "django-filter>=25.1",
    "django-jazzmin>=3.0.1",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.28.0",
    "drf-spectacular-sidecar>=2025.9.1",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "python-decouple>=3.8",
    "redis>=6.4.0",
    "scipy>=1.16.2",
]
